+++ src/lib/components/common/Loader/index.ts
@@ -1,0 +1,3 @@
+import Loader from './Loader.svelte';
+
+export default Loader;
